Modify keymap for OS dvorak remap
When remapping the Model 01 to dvorak using the OS to handle the key remapping, the
handling of =, +, -, [, ], /, \, and | keys don't end up with remappings that are
relative to what is printed on the keycaps (for the function modes).

Additionally, the Model 01's mappings of the "{", "}", "[", and "]" keys don't end
up where they are intended to be used by holding the function key (and as printed on
the keycaps).

Instead, manually juggle the keys on the firmware level, so they end up in the expected
places when the OS does the remapping to dvorak.

This results in:

1) The "/", "?", and "\" are all usable from the same key again.
Pressing "/" in dvorak while holding the function key results in "\".

2) The "-", "_", and "|" are all usable from the same key again.
Pressing "-" key while holding the function key results in "|"

3) Holding the function key while pressing "g", "c", "r", "l" results in
"{", "}", "[", "]", in that order, preserving the convenience location
of squigglies and braces as intended by the Model 01.

4) Since F12 had to be remapped somewhere (to make room for the prior keys being mapped to
the expected places), F12 has been remapped to the bottom right most key on the Model 01
(where "|" would be triggered with the function key if in the stock QWERTY mode.)